requirements.txt
crewai
langchain-community
python-dotenv

